@@ -1,9 +1,8 @@
 package jastzeonic.com.jastzeonictodolist
 
-import android.arch.lifecycle.ViewModelProvider
-import android.arch.lifecycle.ViewModelProviders
-import android.databinding.DataBindingUtil
-import android.support.v7.app.AppCompatActivity
+import androidx.lifecycle.ViewModelProviders
+import androidx.databinding.DataBindingUtil
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import jastzeonic.com.jastzeonictodolist.databinding.ActivitySampleDatabindingBinding
 import jastzeonic.com.jastzeonictodolist.view.model.SampleViewModel
