@@ -1,10 +1,10 @@
 package jastzeonic.com.jastzeonictodolist.model
 
-import android.arch.persistence.room.ColumnInfo
-import android.arch.persistence.room.Entity
-import android.arch.persistence.room.Ignore
-import android.arch.persistence.room.PrimaryKey
-import android.support.annotation.NonNull
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.Ignore
+import androidx.room.PrimaryKey
+import androidx.annotation.NonNull
 
 @Entity(tableName = TodoModel.TABLE_NAME)
 class TodoModel {
