@@ -1,9 +1,9 @@
 package jastzeonic.com.jastzeonictodolist.database
 
-import android.arch.persistence.db.SupportSQLiteDatabase
-import android.arch.persistence.room.Database
-import android.arch.persistence.room.RoomDatabase
-import android.arch.persistence.room.migration.Migration
+import androidx.sqlite.db.SupportSQLiteDatabase
+import androidx.room.Database
+import androidx.room.RoomDatabase
+import androidx.room.migration.Migration
 import android.util.Log
 import jastzeonic.com.jastzeonictodolist.model.TodoModel
 
