@@ -1,10 +1,12 @@
 package jastzeonic.com.jastzeonictodolist.view.model
 
 import android.app.Application
-import android.arch.lifecycle.AndroidViewModel
-import android.arch.lifecycle.MutableLiveData
-import android.databinding.ObservableField
+import androidx.lifecycle.AndroidViewModel
+import androidx.lifecycle.MutableLiveData
+import androidx.databinding.ObservableField
 import android.util.Log
+import io.reactivex.disposables.CompositeDisposable
+import io.reactivex.rxkotlin.addTo
 import jastzeonic.com.jastzeonictodolist.RepositoryProvider
 import jastzeonic.com.jastzeonictodolist.model.TodoModel
 import jastzeonic.com.jastzeonictodolist.model.TodoRepository
@@ -17,12 +19,15 @@ class TodoListViewModel(application: Application) : AndroidViewModel(application
 
     val todoListObservable = ObservableField<List<TodoModel>>()
 
+    var compositeDisposable = CompositeDisposable()
+
 
     private val onDeleteSuccessEvent = MutableLiveData<Boolean>()
 
 
     fun getTodoList() {
         repo.getAll().subscribe({ todoList -> todoListObservable.set(todoList) }, { onError -> Log.d("####", "onError:" + onError.message.toString()) })
+                .addTo(compositeDisposable)
     }
 
 
@@ -31,8 +36,14 @@ class TodoListViewModel(application: Application) : AndroidViewModel(application
         model.id = id
 
         repo.delete(model).subscribe({ onSuccess -> onDeleteSuccessEvent.value = onSuccess > 0 }, {})
+                .addTo(compositeDisposable)
 
     }
 
+
+    override fun onCleared() {
+        super.onCleared()
+        compositeDisposable.clear()
+    }
 
 }
